requirements.txt
annotated-types==0.7.0
anyio==4.9.0
bcrypt==4.3.0
black==25.1.0
cffi==1.17.1
click==8.2.0
colorama==0.4.6
cryptography==44.0.3
fastapi==0.115.12
h11==0.16.0
idna==3.10
joblib==1.5.0
jwcrypto==1.5.6
multipart==1.2.1
mypy_extensions==1.1.0
numpy==2.2.5
packaging==25.0
passlib==1.7.4
pathspec==0.12.1
platformdirs==4.3.8
pycparser==2.22
pydantic==2.11.4
pydantic_core==2.33.2
PyJWT==2.10.1
python-dotenv==1.1.0
python-jwt==4.1.0
python-multipart==0.0.20
scikit-learn==1.6.1
scipy==1.15.3
sniffio==1.3.1
starlette==0.46.2
threadpoolctl==3.6.0
typing-inspection==0.4.0
typing_extensions==4.13.2
uvicorn==0.34.2
